refactor: abstract pagination component
